updated example app
