+++ examples/temporalgrid/package.json
@@ -20,6 +20,7 @@
     "@babel/plugin-proposal-class-properties": "^7.0.0",
     "@babel/preset-env": "^7.9.0",
     "@babel/preset-react": "^7.0.0",
+    "@welldone-software/why-did-you-render": "^5.0.0-alpha.2",
     "babel-loader": "^8.0.0",
     "webpack": "^4.20.0",
     "webpack-cli": "^3.1.2",
